Extract team bulk-upload pipeline into TeamsUploadRunner
Extracts uploadTeams, queueTeamRetry, and uploadTeamImageAttachment from
UploadManager into TeamsUploadRunner under services/upload/ and updates
UploadManager to delegate to it.
